Update code to keep consistecyt with repo standards
