+++ datacatalog/cloud-client/README.md
@@ -24,24 +24,6 @@ the [Google Cloud Client Library for Java][google-cloud-java].
   mvn clean package -DskipTests
 ```
 
-#### Lookup a BigQuery Dataset
-```
-  mvn exec:java -Dexec.mainClass=com.example.datacatalog.LookupEntryBigQueryDataset -Dexec.args="my-project my_dataset"
-```
-
-#### Lookup a BigQuery Table
-```
-  mvn exec:java -Dexec.mainClass=com.example.datacatalog.LookupEntryBigQueryTable -Dexec.args="my-project my_dataset my_table"
-```
-
-#### Lookup a Pub/Sub Topic
-```
-  mvn exec:java -Dexec.mainClass=com.example.datacatalog.LookupEntryPubSubTopic -Dexec.args="my-project my-topic"
-```
-
-#### Lookup by Sql Resource
-Append `-lookupBySqlResource` to command args to lookup by Sql Resource instead of Linked Resource.
-
 #### Testing
 Run the test with Maven.
 ```
