+++ datacatalog/cloud-client/src/main/java/com/example/datacatalog/LookupEntryBigQueryDataset.java
@@ -27,25 +27,21 @@ public class LookupEntryBigQueryDataset {
    *
    * @param projectId The project ID to which the Dataset belongs, e.g. 'my-project'
    * @param datasetId The dataset ID to which the Catalog Entry refers, e.g. 'my_dataset'
-   * @param lookupBySqlResource Indicates whether the lookup should be performed by Sql Resource
-   *     instead of Linked Resource, e.g. 'false'
    */
-  private static void lookupEntryBigQueryDataset(
-      String projectId, String datasetId, boolean lookupBySqlResource) {
+  public static void lookupEntry(String projectId, String datasetId) {
+    // String projectId = "my-project"
+    // String datasetId = "my_dataset"
 
-    LookupEntryRequest request;
+    // Get an entry by the resource name from the source Google Cloud Platform service.
+    String linkedResource =
+        String.format("//bigquery.googleapis.com/projects/%s/datasets/%s", projectId, datasetId);
+    LookupEntryRequest request =
+        LookupEntryRequest.newBuilder().setLinkedResource(linkedResource).build();
 
-    // Construct the Lookup request to be sent by the client.
-    LookupEntryRequest request;
-    if (lookupBySqlResource) {
-    if (lookupBySqlResource) {
-      String sqlResource = String.format("bigquery.dataset.`%s`.`%s`", projectId, datasetId);
-      request = LookupEntryRequest.newBuilder().setSqlResource(sqlResource).build();
-    } else {
-      String linkedResource =
-          String.format("//bigquery.googleapis.com/projects/%s/datasets/%s", projectId, datasetId);
-      request = LookupEntryRequest.newBuilder().setLinkedResource(linkedResource).build();
-    }
+    // Alternatively, lookup by the SQL name of the entry would have the same result:
+    // String sqlResource = String.format("bigquery.dataset.`%s`.`%s`", projectId, datasetId);
+    // LookupEntryRequest request =
+    // LookupEntryRequest.newBuilder().setSqlResource(sqlResource).build();
 
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
@@ -56,19 +52,5 @@ public class LookupEntryBigQueryDataset {
     } catch (Exception e) {
       System.out.print("Error during lookupEntryBigQueryDataset:\n" + e.toString());
     }
-  }
-
-  /**
-   * Command line application to lookup the Data Catalog entry referring to a BigQuery Dataset.
-   * Requires 2 positional args: projectId and datasetId. A third arg is optional:
-   * lookupBySqlResource (when set, the lookup is done by Sql Resource instead of Linked Resource).
-   */
-  public static void main(String... args) {
-
-    String projectId = args[0];
-    String datasetId = args[1];
-    boolean lookupBySqlResource = "-lookupBySqlResource".equals(args[args.length - 1]);
-
-    lookupEntryBigQueryDataset(projectId, datasetId, lookupBySqlResource);
   }
 }
