+++ datacatalog/cloud-client/src/main/java/com/example/datacatalog/QuickStart.java
@@ -1,0 +1,4 @@
+package com.example.datacatalog;
+
+public class QuickStart {
+}
